Created ShapeScore, Score-adhering model
* Takes a `Shapeish` and acts like a `Score` producing a `value: Score.Value`.
* Also updated print output code in main's `inputLines` loop.

Unlike other models I've done, this doesn't wrap a Score, it just adheres to the `Scoreish` interface.  I could've made it wrap a Score model, but really that'd have to be a DynamicScore model and that's a lot of crap for a simple implementation.

I'm still not 100% about how I feel about this— it's breaking the Liskov Substitution Principle, but then again that's sort of by-design with metacosm disallowing subclasses.  But does this still follow it in spirit?  Could a `ShapeScore` instance be passed to anything that expects a `Scoreish` or a hypothetical `DynamicScoreish` without causing any unexpected behavior?  Since the common interface is `Scoreish` which is get-only… I guess, yes?