Upgraded ModelSurrogateWrapper to be able to take a closure
Turned `ModelSurrogateWrapper` into an enum so it can have a `getClosure` mode that'll fetch the `ModelT` on-demand.  Useful when we want the model to be owned/stored by the encompassing class, and we just want the `ModelSurrogateWrapper` to reference it.

Also had to add a 3rd `.unitialized` state, since the owning class can't set a closure reference it`self` until after `super.init()`.